In this commit I was working in the improve of the christine logger capabilites by creating a LoggerManager that lives in the interface singleton and is in charge to give the logger accord to the name given.
Also, there is a new module called options, This module will retain the 
christine options making use of optparse for that.

I have fixed a small regression bug appeared in the search field. 

I test this code and should not break anything but I didn't test it so much
so expect this to fail.

The logger and options still need more love,